USER AUTH STRUCTURE - Created folder UserAuth and moved related form/user auth to it - Edit path on App.js to new file path
